requirements.txt
fastapi>=0.104.1
sqlalchemy>=2.0.23
uvicorn>=0.24.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1
requests>=2.31.0